backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glucotracker-backend"
version = "0.1.0"
description = "FastAPI backend for the Glucotracker personal food diary."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.34",
  "sqlalchemy>=2.0",
  "alembic>=1.14",
  "pydantic>=2.9",
  "pydantic-settings>=2.7",
  "python-multipart>=0.0.20",
  "httpx>=0.28",
  "google-genai>=1.0",
  "pypdf>=6.0",
  "pyyaml>=6.0",
  "pytest>=8.3",
  "pytest-asyncio>=0.25",
  "ruff>=0.8",
]

[tool.setuptools.packages.find]
include = ["glucotracker*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
